Moving all decorators and helper functions out of domainRoutes.py and into the proper files
